pubsub2firestore/requirements.txt
google-cloud-pubsub==2.31.1
google-cloud-firestore==2.20.1
